# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: 41 Tech API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication
  - name: projects
    description: Projects management
  - name: team
    description: Team members management
  - name: cases
    description: Cases management
  - name: technologies
    description: Technologies management

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Admin logout
      responses:
        "200":
          description: Logout successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current session user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      parameters:
        - in: query
          name: featured
          schema:
            type: boolean
        - in: query
          name: status
          schema:
            type: string
      responses:
        "200":
          description: Projects list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectBody"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"

  /projects/{slug}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get project by slug
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateProject
      tags: [projects]
      summary: Update project by slug
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectBody"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete project by slug
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /team:
    get:
      operationId: listTeamMembers
      tags: [team]
      summary: List team members
      responses:
        "200":
          description: Team members list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TeamMember"
    post:
      operationId: createTeamMember
      tags: [team]
      summary: Create team member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTeamMemberBody"
      responses:
        "201":
          description: Team member created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamMember"

  /team/{id}:
    get:
      operationId: getTeamMember
      tags: [team]
      summary: Get team member by id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Team member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamMember"
    put:
      operationId: updateTeamMember
      tags: [team]
      summary: Update team member
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTeamMemberBody"
      responses:
        "200":
          description: Updated team member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamMember"
    delete:
      operationId: deleteTeamMember
      tags: [team]
      summary: Delete team member
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /cases:
    get:
      operationId: listCases
      tags: [cases]
      summary: List cases
      parameters:
        - in: query
          name: public_only
          schema:
            type: boolean
      responses:
        "200":
          description: Cases list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Case"
    post:
      operationId: createCase
      tags: [cases]
      summary: Create case
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCaseBody"
      responses:
        "201":
          description: Case created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Case"

  /cases/{slug}:
    get:
      operationId: getCase
      tags: [cases]
      summary: Get case by slug
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Case
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Case"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateCase
      tags: [cases]
      summary: Update case
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCaseBody"
      responses:
        "200":
          description: Updated case
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Case"
    delete:
      operationId: deleteCase
      tags: [cases]
      summary: Delete case
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /technologies:
    get:
      operationId: listTechnologies
      tags: [technologies]
      summary: List technologies
      responses:
        "200":
          description: Technologies list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Technology"
    post:
      operationId: createTechnology
      tags: [technologies]
      summary: Create technology
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTechnologyBody"
      responses:
        "201":
          description: Technology created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Technology"

  /technologies/{id}:
    put:
      operationId: updateTechnology
      tags: [technologies]
      summary: Update technology
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTechnologyBody"
      responses:
        "200":
          description: Updated technology
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Technology"
    delete:
      operationId: deleteTechnology
      tags: [technologies]
      summary: Delete technology
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /stats/summary:
    get:
      operationId: getStatsSummary
      tags: [health]
      summary: Get summary stats for public display
      responses:
        "200":
          description: Stats summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatsSummary"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success

    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password

    LoginResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/AdminUser"
      required:
        - user

    AdminUser:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
      required:
        - id
        - name
        - email
        - role

    Project:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        shortDescription:
          type: string
        fullDescription:
          type: string
          nullable: true
        problem:
          type: string
          nullable: true
        solution:
          type: string
          nullable: true
        result:
          type: string
          nullable: true
        coverImageUrl:
          type: string
          nullable: true
        demoUrl:
          type: string
          nullable: true
        repositoryUrl:
          type: string
          nullable: true
        status:
          type: string
        featured:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - title
        - slug
        - shortDescription
        - status
        - featured
        - createdAt
        - updatedAt

    CreateProjectBody:
      type: object
      properties:
        title:
          type: string
        slug:
          type: string
        shortDescription:
          type: string
        fullDescription:
          type: string
          nullable: true
        problem:
          type: string
          nullable: true
        solution:
          type: string
          nullable: true
        result:
          type: string
          nullable: true
        coverImageUrl:
          type: string
          nullable: true
        demoUrl:
          type: string
          nullable: true
        repositoryUrl:
          type: string
          nullable: true
        status:
          type: string
        featured:
          type: boolean
      required:
        - title
        - slug
        - shortDescription
        - status
        - featured

    TeamMember:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        roleTitle:
          type: string
        bio:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        linkedinUrl:
          type: string
          nullable: true
        githubUrl:
          type: string
          nullable: true
        sortOrder:
          type: integer
        isActive:
          type: boolean
        skills:
          type: array
          items:
            type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - roleTitle
        - sortOrder
        - isActive
        - skills
        - createdAt
        - updatedAt

    CreateTeamMemberBody:
      type: object
      properties:
        name:
          type: string
        roleTitle:
          type: string
        bio:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        linkedinUrl:
          type: string
          nullable: true
        githubUrl:
          type: string
          nullable: true
        sortOrder:
          type: integer
        isActive:
          type: boolean
        skills:
          type: array
          items:
            type: string
      required:
        - name
        - roleTitle
        - sortOrder
        - isActive
        - skills

    Case:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        clientSegment:
          type: string
          nullable: true
        problem:
          type: string
        solution:
          type: string
        result:
          type: string
        metricsSummary:
          type: string
          nullable: true
        isPublic:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - title
        - slug
        - problem
        - solution
        - result
        - isPublic
        - createdAt
        - updatedAt

    CreateCaseBody:
      type: object
      properties:
        title:
          type: string
        slug:
          type: string
        clientSegment:
          type: string
          nullable: true
        problem:
          type: string
        solution:
          type: string
        result:
          type: string
        metricsSummary:
          type: string
          nullable: true
        isPublic:
          type: boolean
      required:
        - title
        - slug
        - problem
        - solution
        - result
        - isPublic

    Technology:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        iconUrl:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - createdAt
        - updatedAt

    CreateTechnologyBody:
      type: object
      properties:
        name:
          type: string
        iconUrl:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
      required:
        - name

    StatsSummary:
      type: object
      properties:
        totalProjects:
          type: integer
        featuredProjects:
          type: integer
        totalTeamMembers:
          type: integer
        totalCases:
          type: integer
        totalTechnologies:
          type: integer
      required:
        - totalProjects
        - featuredProjects
        - totalTeamMembers
        - totalCases
        - totalTechnologies