# lib/api-zod/src/generated/types/case.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 41 Tech API
 * OpenAPI spec version: 0.1.0
 */

export interface Case {
  id: number;
  title: string;
  slug: string;
  clientSegment?: string | null;
  problem: string;
  solution: string;
  result: string;
  metricsSummary?: string | null;
  isPublic: boolean;
  createdAt: string;
  updatedAt: string;
}