# lib/api-zod/src/generated/types/createTeamMemberBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 41 Tech API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateTeamMemberBody {
  name: string;
  roleTitle: string;
  bio?: string | null;
  avatarUrl?: string | null;
  linkedinUrl?: string | null;
  githubUrl?: string | null;
  sortOrder: number;
  isActive: boolean;
  skills: string[];
}