# lib/api-zod/src/generated/types/statsSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 41 Tech API
 * OpenAPI spec version: 0.1.0
 */

export interface StatsSummary {
  totalProjects: number;
  featuredProjects: number;
  totalTeamMembers: number;
  totalCases: number;
  totalTechnologies: number;
}