# lib/api-zod/src/generated/types/technology.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 41 Tech API
 * OpenAPI spec version: 0.1.0
 */

export interface Technology {
  id: number;
  name: string;
  iconUrl?: string | null;
  category?: string | null;
  createdAt: string;
  updatedAt: string;
}